Only start the client in the test driver
If user tests need any additional machines online, they need to bring
them up themselves (explicit is better than implicit and all that).
As a bonus, tests now properly wait for the server to start listening
instead of hoping it was brought up in the background during the
client setup.